+++ editor/src/index.scss
@@ -9,17 +9,17 @@
 body {
   margin: 0;
   padding: 0;
-  font-family: Inter, sans-serif;
+  font-family: "DM Sans", sans-serif;
 }
 
 .bp5-button,
 .bp5-control,
 .bp5-input {
-  font-family: Inter, sans-serif;
+  font-family: "DM Sans", sans-serif;
 }
 
 button {
-  font-family: Inter, sans-serif;
+  font-family: "DM Sans", sans-serif;
 }
 
 .home-page {
