+++ editor/src/integrated-flow-manager/App.scss
@@ -11,7 +11,7 @@
   flex: 1;
   display: flex;
   flex-direction: column;
-  font-family: Inter, sans-serif;
+  font-family: "DM Sans", sans-serif;
   height: 100vh;
 
   &.embedded {
