+++ flow-editor/src/index.scss
@@ -1,4 +1,4 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700");
+@import url("https://fonts.googleapis.com/css2?family=DM+Sans:wght@300;400;500;600;700");
 
 @import "~@blueprintjs/core/lib/css/blueprint.css";
 @import "~@blueprintjs/select/lib/css/blueprint-select.css";
@@ -13,13 +13,13 @@
 body {
   margin: 0;
   padding: 0;
-  font-family: Inter, sans-serif;
+  font-family: "DM Sans", sans-serif;
 }
 
 .bp5-button,
 .bp5-control,
 .bp5-input {
-  font-family: Inter, sans-serif;
+  font-family: "DM Sans", sans-serif;
 }
 
 .home-page {
