+++ flow-editor/src/lib/common.scss
@@ -4,7 +4,7 @@
 $selection-color: #007ec6;
 $selection-text-color: #e9e9e9;
 
-$font-family: Inter, sans-serif;
+$font-family: "DM Sans", sans-serif;
 
 $shadow-a: 0px 0px 10px 2px rgba(48, 48, 48, 0.15);
