+++ flow-editor/src/visual-node-editor/NodesLibrary/NodesLibrary.scss
@@ -11,7 +11,7 @@
   position: absolute;
   right: 0px;
   top: 15%;
-  font-size: 12px;
+  font-size: 14px;
   transform-origin: top right;
   border-radius: 5px;
   max-height: 70%;
@@ -125,7 +125,7 @@
     user-select: none;
     font-weight: 600;
     margin-bottom: 3px;
-    font-size: 11px;
+    font-size: 13px;
   }
 
   .group-items {
@@ -139,7 +139,7 @@
 
   .group-item {
     font-weight: 500;
-    font-size: 11px;
+    font-size: 13px;
     cursor: pointer;
     user-select: none;
     border-radius: 6px;
