+++ flow-editor/src/visual-node-editor/connection-view/ConnectionViewPath/ConnectionViewPath.tsx
@@ -2,6 +2,7 @@ import { Pos } from "@flyde/core";
 import classNames from "classnames";
 import React, { forwardRef } from "react";
 import { calcBezierPath } from "../bezier";
+import { ContextMenu, Menu, MenuItem } from "@blueprintjs/core";
 
 export interface ConnectionViewPathProps {
   from: Pos;
@@ -12,12 +13,21 @@ export interface ConnectionViewPathProps {
   onMouseLeave?: (e: React.MouseEvent<any, MouseEvent>) => void;
   zoom: number;
   dashed?: boolean;
+  onDelete?: () => void;
 }
 
 export const ConnectionViewPath: React.FC<ConnectionViewPathProps> = forwardRef(
   (props, ref) => {
-    const { from, to, className, zoom, dashed, onMouseEnter, onMouseLeave } =
-      props;
+    const {
+      from,
+      to,
+      className,
+      zoom,
+      dashed,
+      onMouseEnter,
+      onMouseLeave,
+      onDelete,
+    } = props;
     const { x: x1, y: y1 } = from;
     const { x: x2, y: y2 } = to;
 
@@ -51,15 +61,33 @@ export const ConnectionViewPath: React.FC<ConnectionViewPathProps> = forwardRef(
       />
     );
 
+    const contextMenuContent = (
+      <Menu>
+        <MenuItem text="Delete connection" intent="danger" onClick={onDelete} />
+      </Menu>
+    );
+
     return (
-      <>
-        {path}
-        {pathProximityMask}
-        <linearGradient id="gradient" x1="0%" y1="0%" x2="100%" y2="0%">
-          <stop offset="0%" />
-          <stop offset="100%" />
-        </linearGradient>
-      </>
+      <ContextMenu content={contextMenuContent}>
+        {(ctxMenuProps) => (
+          <g
+            className={classNames(
+              "connection-view-path",
+              ctxMenuProps.className
+            )}
+            onContextMenu={ctxMenuProps.onContextMenu as any}
+            ref={ctxMenuProps.ref as any}
+          >
+            {ctxMenuProps.popover}
+            {path}
+            {pathProximityMask}
+            <linearGradient id="gradient" x1="0%" y1="0%" x2="100%" y2="0%">
+              <stop offset="0%" />
+              <stop offset="100%" />
+            </linearGradient>
+          </g>
+        )}
+      </ContextMenu>
     );
   }
 );
